Changed design/implementation of prefix-types to not generate a `*_WithMetadata` struct.
Changed layout checking to always run in the binary,
passing a function pointer to that function to libraries.

Moved examples to example folder.

Removed LibraryTrait/ModuleTrait in favor of `trait RootModule`
implemented by the root module.

Added example of json serialization/deserialization in the example 0.

Added tests for `#[sabi(..)]` attributes related to prefix-types.