Added testing framework for abi_stable_derive that uses config files for tests.
The testing framework uses `.ron` (Rusty object notation) files to construct tests,
which test that a proc-macro returns either and error or not,
and that (optionally) checks the text that the return value contains.

Added some depenencies for the testing framework.

Defined a RegexWrapper type to deserialize regexes and use them as keys.