Added range checking to abi_stable_derive,with tests.
abi_stable_derive now returns errors if it would overflow the constants
it generates.

Rewrote a lot of code to use `syn::Ident` instead of strings,
to get the span for error messages.

Made LifetimeIndex be represented as a usize in abi_stable_derive,
so that it doesn't overflow while visiting a function pointer type.

Added TypeLayoutIndex to represent the integer values