First prefix-type size reduction.
Moved strings and lifetime indices to MonoSharedVars,which is inside MonoTypeLayout.

Added a reference to MonoTypeLayout in PTStructLayout,
removing PTStructLayoutParams and some other fields.

Moved PrefixConditionality from the trait to PTStructLayout.

Added `RSlice::transmute_ref` to transmute RSLices in const contexts.

Added `_sabi_type_layouts` macro to construct the list of nested
type layouts in TypeLayout.

Emiting StartLen as a u32 in the derive macro.