Rewrote how prefix types work, MEGACOMMIT.
Only made sure that abi_stable and abi_stable_derive compile,and the tests pass.
`abi_stable::library` is commented out, that will be fixed in a later commit.

Declared the PrefixRef type, a non-nullable pointer for all prefix types.
This type replaces all instances of the `&Prefix` and `StaticRef<Prefix>`.

Now prefix don't work using references to the prefix,
instead there's a `PrefixRef` newtype over a NonNull,
which is the only way to access the prefix.
And every type has a `DerivingType_Ref` type (newtype around a `PrefixRef`),
which defines the accessor methods for the fields in `DerivingType`

Now the `#[sabi(kind(Prefix))]` attribute takes 2 optional arguments:
- `prefix_ref`(defaults to `<DerivingType>_Ref`):
    The name of the generated pointer type, with accessor methods for
    the fields of the deriving type (some of which are conditional).
- `prefix_fields`(defaults to `<DerivingType>_Prefix`):
    The name of the generated prefix struct,
    with only the fields up to(and including) the #[sabi(last_prefix_field)]` attribute.

Removed the `*Val` suffix for types deriving StableAbi with a prefix kind.

Updated the tests to compile and pass.

Updated the doctests to pass, and changed some tests to
use a `const` instead of leaking to construct the pointer to prefix.

Replaced virtually all uses of StaticRef with PrefixRef (or newtypes that wrap it),
since all the replaced uses were for prefix types.
Changed StaticRef to use NonNull, so that Option<StaticRef> is ffi-safe.
Updated the StaticRef example to be for a non-extensible vtable.

Removed the Kind associated type from StableAbi.

Split off SharedStableAbi trait from StableAbi, now there's PrefixStableAbi and StableAbi,
both of which only have StaticEquivalent_ in common.

Added NonNullPointer for later use in generalizing LateStaticRef to
support any type wrapping a non-nullable pointer
(including the `DerivingType_Ref` types).

Replaced WithMetadataFor with the new PrefixMetadata,
with the type layout for prefix field error messages.
Defined accessor methods for PrefixMetadata's fields.
And storing PrefixMetadata as the first field of WithMetadata_.

Defined WithMetadata_ methods for returning a `PrefixRef`,removing the StaticRef ones.

Removed PrefixTypeTrait::into_with_metadata method,added PrefixFields and PrefixRef assoc types.

Made the type parameters in the vtable of `#[sabi_trait]` trait objects covariant.

The <DerivingType>_Ref type implements the Copy,Clone,GetStaticEquivalent_,StableAbi,
NonNullPointer traits

Changed how prefix type accessor methods work,
fields after the suffix are accessed by offsetting a raw pointer to the offset of the field.
The field offsets are computed at compile-time,using a function from the `repr_orrset` crate,

Added `repr_offset` as a dependency of `abi_stable_derive`