Added test crates for loading errors,bumped MSRV,refactored some code.
Added crates in `/testing/1 - loading errors/` to test the errors that happen while
loading a dynamic library, starting with abi incompatibility errors.

Moved the `compute_library_path` function in a bunch of the "user" crates to the newly created
`abi_stable::library::development_utils` module,
made the function take the path to the target and the RootModule type as parameters.

Moved the `testing/*_0` crates to the new `testing/0/` directory.

Updated the travis configuration (will probably have to fix it when CI runs),

Bumping the minimum supported rust version, update the readme and in CI tests.