Fixed miri testability of #[sabi_trait] trait objects.
Defined the RMut type for use as the self parameter in the implementation of the trait objects.

Changed the implementation of the trait objects to use
RRef for &self, RMut for &mut self, and *mut () for self.

Added sabi_as_rref and sabi_as_rmut methods to RObject and DynTrait to get RRef and RMut.