Made layout tests runnable in miri.
Calling the `miri_static_root` function to tell miri that the memory I leak intentionally
is actually used.

Extracted a more minimal version of many layout tests,to run in a reasonable amount of time in miri.
This means that there's versions of those functions for miri and not.

Updated travis config to run miri tests, hopefully this won't run out of RAM.