Replaced type erasure of &/&mut with RRef/RMut
Changed & and &mut to RRef and RMut in functions taking/returning
type-erased references because casting `&ErasedObject` to `&T` is
incompatible with the Stacked Borrows model (and miri).

Added AsPtr and AsMutPtr traits. implemented for all erasable pointer types.

Done to `CanTransmuteElement`:
    - Added `transmute_element_` method
    - Changed TransmutedPtr bound to `AsPtr<Target = T>`

Replaced all Deref and DerefMut bounds in trait object types with AsPtr and AsMutPtr

Changed CanTransmuteElement impls for references to return RRef and RMut

Fixed the soundness bug where pointers to serde_json::RawValue could change the
order of the length and data pointers when transmuted to point to str.

Added these methods to AsMut:
    - transmute_ref_mut
    - into_rref
    - as_rref