Removed `#[sabi_trait]` vtable erasure hack
This hack was necessary to construct trait objects in consts,
but now that trait bounds are allowed in const fns it is no longer needed.

Fixed a bug where trait method with `&'a self`
(where `'a` is a parameter of the trait)
could be called when the equivalent Rust trait object could not.

Added UI test for the bugfix